비밀번호 재설정 기능 구현

+++ src/main/java/io/cavia/trader/module/member/mypage/service/MyPageService.java
@@ -16,7 +16,7 @@ public interface MyPageService {
 
     public boolean validatePassword(Long id, String password);
 
-    public int changePassword(int id, String password, LocalDateTime passwordUpdatedAt);
+    public int changePassword(Long id, String password, LocalDateTime passwordUpdatedAt);
 
     public int resetCash(Long id);
